requirements.txt
flask
torch
torchvision
diffusers
transformers
accelerate
pillow
controlnet-aux
opencv-python
numpy
timm
